+++ EU4toV3/EU4ToV3.vcxproj
@@ -144,9 +144,10 @@ call save_commit_id.bat</Command>
     <ClCompile Include="Source\Output\outStates\outSubState.cpp" />
     <ClCompile Include="Source\Output\outWorld.cpp" />
     <ClCompile Include="Source\V3World\ClayManager\ClayManager.cpp" />
-    <ClCompile Include="Source\V3World\ClayManager\PopLoader\Pop.cpp" />
     <ClCompile Include="Source\V3World\ClayManager\PopLoader\PopEntry.cpp" />
     <ClCompile Include="Source\V3World\ClayManager\PopLoader\PopLoader.cpp" />
+    <ClCompile Include="Source\V3World\ClayManager\PopLoader\Pops\StatePops.cpp" />
+    <ClCompile Include="Source\V3World\ClayManager\PopLoader\Pops\SubStatePops.cpp" />
     <ClCompile Include="Source\V3World\ClayManager\PopLoader\StateEntries.cpp" />
     <ClCompile Include="Source\V3World\ClayManager\PopLoader\SubStatePopEntries.cpp" />
     <ClCompile Include="Source\V3World\ClayManager\PopLoader\SubStateEntries.cpp" />
@@ -263,9 +264,11 @@ call save_commit_id.bat</Command>
     <ClInclude Include="Source\Output\outWorld.h" />
     <ClInclude Include="Source\V3World\ClayManager\Chunk.h" />
     <ClInclude Include="Source\V3World\ClayManager\ClayManager.h" />
-    <ClInclude Include="Source\V3World\ClayManager\PopLoader\Pop.h" />
     <ClInclude Include="Source\V3World\ClayManager\PopLoader\PopEntry.h" />
     <ClInclude Include="Source\V3World\ClayManager\PopLoader\PopLoader.h" />
+    <ClInclude Include="Source\V3World\ClayManager\PopLoader\Pops\Pop.h" />
+    <ClInclude Include="Source\V3World\ClayManager\PopLoader\Pops\StatePops.h" />
+    <ClInclude Include="Source\V3World\ClayManager\PopLoader\Pops\SubStatePops.h" />
     <ClInclude Include="Source\V3World\ClayManager\PopLoader\StateEntries.h" />
     <ClInclude Include="Source\V3World\ClayManager\PopLoader\SubStatePopEntries.h" />
     <ClInclude Include="Source\V3World\ClayManager\PopLoader\SubStateEntries.h" />
