+++ EU4toV3/Source/V3World/ClayManager/PopLoader/PopLoader.h
@@ -1,7 +1,7 @@
 #ifndef POP_LOADER_H
 #define POP_LOADER_H
 #include "Parser.h"
-#include "Pop.h"
+#include "Pops/StatePops.h"
 
 namespace V3
 {
