+++ EU4toV3/Source/V3World/ClayManager/PopLoader/Pops/SubStatePops.cpp
@@ -1,0 +1,9 @@
+#include "SubStatePops.h"
+#include <numeric>
+
+int V3::SubStatePops::getPopCount() const
+{
+	return std::accumulate(pops.begin(), pops.end(), 0, [](int sum, const auto& pop) {
+		return sum + pop.getSize();
+	});
+}
