+++ EU4toV3/Source/V3World/ClayManager/PopLoader/StateEntries.h
@@ -1,7 +1,7 @@
 #ifndef STATE_ENTRIES_H
 #define STATE_ENTRIES_H
 #include "Parser.h"
-#include "SubStatePopEntries.h"
+#include "Pops/StatePops.h"
 
 namespace V3
 {
