+++ EU4toV3/Source/V3World/ClayManager/PopLoader/SubStatePopEntries.h
@@ -1,7 +1,7 @@
 #ifndef SUBSTATE_POP_ENTRIES_H
 #define SUBSTATE_POP_ENTRIES_H
 #include "Parser.h"
-#include "PopEntry.h"
+#include "Pops/Pop.h"
 
 namespace V3
 {
