+++ EU4toV3Tests/EU4ToV3Tests.vcxproj
@@ -144,9 +144,10 @@
     <ClCompile Include="..\EU4toV3\Source\Mappers\SuperGroupMapper\SuperGroupMapper.cpp" />
     <ClCompile Include="..\EU4toV3\Source\Mappers\SuperGroupMapper\SuperRegionParser.cpp" />
     <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\ClayManager.cpp" />
-    <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\Pop.cpp" />
     <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\PopEntry.cpp" />
     <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\PopLoader.cpp" />
+    <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\Pops\StatePops.cpp" />
+    <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\Pops\SubStatePops.cpp" />
     <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\StateEntries.cpp" />
     <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\SubStateEntries.cpp" />
     <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\SubStatePopEntries.cpp" />
