+++ EU4toV3Tests/EU4ToV3Tests.vcxproj.filters
@@ -532,8 +532,11 @@
     <ClCompile Include="V3WorldTests\ClayManagerTests\PopLoaderTests\SubStatePopEntriesTests.cpp">
       <Filter>V3WorldTests\ClayManagerTests\PopLoaderTests</Filter>
     </ClCompile>
-    <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\Pop.cpp">
-      <Filter>ConverterFiles\V3World\ClayManager\PopLoader</Filter>
+    <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\Pops\StatePops.cpp">
+      <Filter>ConverterFiles\V3World\ClayManager\PopLoader\Pops</Filter>
+    </ClCompile>
+    <ClCompile Include="..\EU4toV3\Source\V3World\ClayManager\PopLoader\Pops\SubStatePops.cpp">
+      <Filter>ConverterFiles\V3World\ClayManager\PopLoader\Pops</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -869,6 +872,9 @@
     </Filter>
     <Filter Include="TestFiles\vic3installation\game\common\history\pops">
       <UniqueIdentifier>{cfcab57a-62a5-4b91-ad41-791d3145d89e}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="ConverterFiles\V3World\ClayManager\PopLoader\Pops">
+      <UniqueIdentifier>{e6581bfa-a132-41e6-9d15-9a7b438ec623}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
